Moved logic to CoreNativeManager where possible
