+++ src/Whim/Native/CoreNativeManager.cs
@@ -8,11 +8,20 @@ using Windows.Win32.UI.Accessibility;
 using Windows.Win32.UI.Shell.Common;
 using Windows.Win32.UI.WindowsAndMessaging;
 using Windows.Win32.UI.Input.KeyboardAndMouse;
+using System.Collections.Generic;
+using Windows.Win32.Graphics.Dwm;
 
 namespace Whim;
 
 internal class CoreNativeManager : ICoreNativeManager
 {
+	private readonly IConfigContext _configContext;
+
+	public CoreNativeManager(IConfigContext configContext)
+	{
+		_configContext = configContext;
+	}
+
 	/// <inheritdoc/>
 	public UnhookWinEventSafeHandle SetWinEventHook(uint eventMin, uint eventMax, WINEVENTPROC lpfnWinEventProc) =>
 		PInvoke.SetWinEventHook(eventMin, eventMax, null, lpfnWinEventProc, 0, 0, PInvoke.WINEVENT_OUTOFCONTEXT);
@@ -113,5 +122,150 @@ internal class CoreNativeManager : ICoreNativeManager
 		}
 		lpdwProcessId = pid;
 		return creator;
+	}
+
+	private const int _bufferCapacity = 255;
+
+	/// <inheritdoc/>
+	public string GetWindowText(HWND hwnd)
+	{
+		unsafe
+		{
+			fixed (char* buffer = new char[_bufferCapacity])
+			{
+				int length = PInvoke.GetWindowText(hwnd, buffer, _bufferCapacity + 1);
+				return length > 0 ? new string(buffer) : "";
+			}
+		}
+	}
+
+	private const string _splashClassName = "MsoSplash";
+
+	/// <inheritdoc/>
+	public bool IsSplashScreen(HWND hwnd)
+	{
+		string classname = _configContext.NativeManager.GetClassName(hwnd);
+		if (classname.Length == 0)
+		{
+			return false;
+		}
+
+		return classname == _splashClassName;
+	}
+
+	/// <inheritdoc/>
+	public IEnumerable<HWND> GetAllWindows()
+	{
+		List<HWND> windows = new();
+
+		PInvoke.EnumWindows(
+			(handle, param) =>
+			{
+				windows.Add(handle);
+				return (BOOL)true;
+			},
+			0
+		);
+
+		return windows;
+	}
+
+	/// <inheritdoc/>
+	public bool IsStandardWindow(HWND hwnd)
+	{
+		if (PInvoke.GetAncestor(hwnd, GET_ANCESTOR_FLAGS.GA_ROOT) != hwnd || !PInvoke.IsWindowVisible(hwnd))
+		{
+			return false;
+		}
+
+		uint style = (uint)PInvoke.GetWindowLong(hwnd, WINDOW_LONG_PTR_INDEX.GWL_STYLE);
+		uint exStyle = (uint)PInvoke.GetWindowLong(hwnd, WINDOW_LONG_PTR_INDEX.GWL_EXSTYLE);
+
+		// WS_POPUP need to have a border or minimize/maximize buttons,
+		// otherwise the window is "not interesting"
+		if (
+			(style & (uint)WINDOW_STYLE.WS_POPUP) == (uint)WINDOW_STYLE.WS_POPUP
+			&& (style & (uint)WINDOW_STYLE.WS_THICKFRAME) == 0
+			&& (style & (uint)WINDOW_STYLE.WS_MINIMIZEBOX) == 0
+			&& (style & (uint)WINDOW_STYLE.WS_MAXIMIZEBOX) == 0
+		)
+		{
+			return false;
+		}
+		if (
+			(style & (uint)WINDOW_STYLE.WS_CHILD) == (uint)WINDOW_STYLE.WS_CHILD
+			|| (style & (uint)WINDOW_STYLE.WS_DISABLED) == (uint)WINDOW_STYLE.WS_DISABLED
+			|| (exStyle & (uint)WINDOW_EX_STYLE.WS_EX_TOOLWINDOW) == (uint)WINDOW_EX_STYLE.WS_EX_TOOLWINDOW
+			|| (exStyle & (uint)WINDOW_EX_STYLE.WS_EX_NOACTIVATE) == (uint)WINDOW_EX_STYLE.WS_EX_NOACTIVATE
+		)
+		{
+			return false;
+		}
+
+		string className = _configContext.NativeManager.GetClassName(hwnd);
+		if (IsSystemWindow(hwnd, className))
+		{
+			return false;
+		}
+
+		return true;
+	}
+
+	private readonly HashSet<string> _systemClasses =
+		new() { "SysListView32", "WorkerW", "Shell_TrayWnd", "Shell_SecondaryTrayWnd", "Progman" };
+
+	/// <inheritdoc/>
+	public bool IsSystemWindow(HWND hwnd, string className)
+	{
+		if (hwnd == PInvoke.GetDesktopWindow() || hwnd == PInvoke.GetShellWindow())
+		{
+			return true;
+		}
+
+		if (_systemClasses.Contains(className))
+		{
+			return true;
+		}
+
+		return false;
+	}
+
+	/// <inheritdoc/>
+	public bool HasNoVisibleOwner(HWND hwnd)
+	{
+		HWND owner = PInvoke.GetWindow(hwnd, GET_WINDOW_CMD.GW_OWNER);
+
+		// The following warning was disabled, since GetWindow can return null, per
+		// https://docs.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-getwindow
+#pragma warning disable CS8073 // The result of the expression is always the same since a value of this type is never equal to 'null'
+		if (owner == null)
+		{
+			return true; // There is no owner at all
+		}
+#pragma warning restore CS8073 // The result of the expression is always the same since a value of this type is never equal to 'null'
+
+		if (!PInvoke.IsWindowVisible(owner))
+		{
+			return true; // Owner is invisible
+		}
+
+		if (!PInvoke.GetWindowRect(owner, out RECT rect))
+		{
+			return false; // Could not get the rect, return true (and filter out the window) just in case
+		}
+
+		// It is enough that the window is zero-sized in one dimension only.
+		return rect.top == rect.bottom || rect.left == rect.right;
+	}
+
+	/// <inheritdoc/>
+	public bool IsCloakedWindow(HWND hwnd)
+	{
+		unsafe
+		{
+			int cloaked;
+			HRESULT res = PInvoke.DwmGetWindowAttribute(hwnd, DWMWINDOWATTRIBUTE.DWMWA_CLOAKED, &cloaked, sizeof(int));
+			return cloaked != 0 || res.Failed;
+		}
 	}
 }
