+++ src/Whim/Native/ICoreNativeManager.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Drawing;
 using System.Runtime.InteropServices;
 using Windows.Win32;
@@ -15,8 +16,6 @@ namespace Whim;
 /// </summary>
 internal interface ICoreNativeManager
 {
-	public static ICoreNativeManager Instance { get; set; } = new CoreNativeManager();
-
 	/// <summary>
 	/// Set the <see cref="PInvoke.SetWinEventHook(uint, uint, SafeHandle, WINEVENTPROC, uint, uint, uint)"/> <br/>
 	///
@@ -316,4 +315,56 @@ internal interface ICoreNativeManager
 	/// <param name="lpdwProcessId"></param>
 	/// <returns></returns>
 	public uint GetWindowThreadProcessId(HWND hWnd, out uint lpdwProcessId);
+
+	/// <summary>
+	/// Safe wrapper around <see cref="PInvoke.GetWindowText"/>.
+	/// </summary>
+	/// <param name="hwnd"></param>
+	/// <returns></returns>
+	public string GetWindowText(HWND hwnd);
+
+	/// <summary>
+	/// Returns <see langword="true"/> if the window is a splash window.
+	/// </summary>
+	/// <param name="hwnd"></param>
+	/// <returns></returns>
+	public bool IsSplashScreen(HWND hwnd);
+
+	/// <summary>
+	/// Enumerates over all the <see cref="HWND"/> of all the top-level windows.
+	/// </summary>
+	/// <returns></returns>
+	public IEnumerable<HWND> GetAllWindows();
+
+	/// <summary>
+	/// Returns <see langword="true"/> when the window is a cloaked window.
+	/// For example, and empty <c>ApplicationFrameWindow</c>.
+	/// For more, see https://social.msdn.microsoft.com/Forums/vstudio/en-US/f8341376-6015-4796-8273-31e0be91da62/difference-between-actually-visible-and-not-visiblewhich-are-there-but-we-cant-see-windows-of?forum=vcgeneral
+	/// </summary>
+	/// <param name="hwnd"></param>
+	/// <returns></returns>
+	public bool IsCloakedWindow(HWND hwnd);
+
+	/// <summary>
+	/// Returns <see langword="true"/> if the window is a standard window.
+	/// Based on https://github.com/microsoft/PowerToys/blob/fa81968dbb58a0697c45335a8f453e5794852348/src/modules/fancyzones/FancyZonesLib/FancyZones.cpp#L381
+	/// </summary>
+	/// <param name="hwnd"></param>
+	/// <returns></returns>
+	public bool IsStandardWindow(HWND hwnd);
+
+	/// <summary>
+	/// Returns <see langword="true"/> when the window has no visible owner.
+	/// </summary>
+	/// <param name="hwnd"></param>
+	/// <returns></returns>
+	public bool HasNoVisibleOwner(HWND hwnd);
+
+	/// <summary>
+	/// Returns <see langword="true"/> if the window is a system window.
+	/// </summary>
+	/// <param name="hwnd"></param>
+	/// <param name="className">The window's class name.</param>
+	/// <returns></returns>
+	public bool IsSystemWindow(HWND hwnd, string className);
 }
